@@ -15,16 +15,24 @@
     "jsonwebtoken": "^8.5.1",
     "koa": "^2.10.0",
     "koa-bodyparser": "^4.2.1",
-    "koa-json": "^2.0.2",
     "koa-jwt": "^3.6.0",
     "koa-logger": "^3.2.1",
     "koa-router": "^7.4.0",
     "mysql2": "^1.7.0",
-    "sequelize": "^5.21.1"
+    "reflect-metadata": "^0.1.13",
+    "typeorm": "^0.2.20"
   },
   "devDependencies": {
-    "eslint": "^6.5.1",
+    "@types/bcrypt": "^3.0.0",
+    "@types/jsonwebtoken": "^8.3.5",
+    "@types/koa-logger": "^3.1.1",
+    "@types/koa-router": "^7.0.42",
+    "@typescript-eslint/eslint-plugin": "^2.6.0",
+    "@typescript-eslint/parser": "^2.6.0",
+    "eslint": "^6.6.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.18.2"
+    "eslint-plugin-import": "^2.18.2",
+    "ts-node": "^8.4.1",
+    "typescript": "^3.6.4"
   }
 }
